feat: Rename package from proxmox-utilities to proxbox
This commit renames the entire project from "proxmox-utilities" to "proxbox"
for better branding and memorability.

Changes:
- Renamed PyPI package from proxmox-utilities to proxbox
- Renamed Python module from proxmox_utilities to proxbox
- Added new CLI entry point: proxbox (proxmox still available for backward compatibility)
- Updated all imports across source files and tests
- Updated README.md with new branding and examples
- Updated CHANGELOG.md with migration information
- Updated CLAUDE.md project context
- Updated all documentation files in docs/
- Updated mkdocs.yml configuration

BREAKING CHANGE: Package name changed from proxmox-utilities to proxbox.
Users will need to install the new package name: pip install proxbox
