+++ src/proxbox/exceptions.py
